requirements.txt
smolagents
requests
openai
PyGithub
gitingest
python-dotenv